server/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
werkzeug==3.0.3
mysql-connector-python==8.4.0
# mysql==0.0.3
requests==2.31.0
selectolax==0.3.21